requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
langchain==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-community==0.3.0
chromadb==0.5.0
python-dotenv==1.0.1
loguru==0.7.2
sentence-transformers==3.3.1